+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Changed
-- Move more tests to the new testing framework (#653)
+- Move more tests to the new testing framework (#653, #655)
 - Namespace the legacy tests (#599)
 
 ### Deprecated